bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["iizhukov <iizhukov@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.5.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "runbot:main"
server = "runserver:main"
